Move hostedcluster manifests package up one level
