+++ hypershift-operator/controllers/manifests/autoscaler/manifests.go
