+++ hypershift-operator/controllers/manifests/clusterapi/manifests.go
