+++ hypershift-operator/controllers/manifests/controlplaneoperator/manifests.go
