+++ hypershift-operator/controllers/manifests/manifests.go
