+++ hypershift-operator/controllers/nodepool/nodepool_controller.go
@@ -10,8 +10,8 @@ import (
 	"github.com/go-logr/logr"
 	hyperv1 "github.com/openshift/hypershift/api/v1alpha1"
 	"github.com/openshift/hypershift/control-plane-operator/releaseinfo"
-	"github.com/openshift/hypershift/hypershift-operator/controllers/hostedcluster/manifests"
 	"github.com/openshift/hypershift/hypershift-operator/controllers/machineimage"
+	"github.com/openshift/hypershift/hypershift-operator/controllers/manifests"
 	hyperutil "github.com/openshift/hypershift/hypershift-operator/controllers/util"
 	capiv1 "github.com/openshift/hypershift/thirdparty/clusterapi/api/v1alpha4"
 	"github.com/openshift/hypershift/thirdparty/clusterapi/util"
